requirements.txt
fastapi
uvicorn
httpx
google-generativeai
motor
payos